Per feedback on PR, moving location of ner tests into tests/nlu/base/test_evaluation.py
