+++ tests/nlu/base/test_evaluation.py
@@ -18,6 +18,7 @@ from rasa.nlu.test import does_token_cross_borders
 from rasa.nlu.test import align_entity_predictions
 from rasa.nlu.test import determine_intersection
 from rasa.nlu.test import determine_token_labels
+from rasa_nlu.test import collect_ner_results
 from rasa.nlu.config import RasaNLUModelConfig
 from rasa.nlu.tokenizers import Token
 from rasa.nlu import utils
@@ -420,3 +421,39 @@ def test_label_replacement():
     target_labels = ["no_entity", "location"]
     assert substitute_labels(original_labels,
                              "O", "no_entity") == target_labels
+
+
+# read generated json file and return the number of tp, fp and fn
+def analyze_ner_file(ner_filename):
+    with open(ner_filename, 'r', encoding='utf-8') as fs:
+        data = json.load(fs)
+    tp_count = len(data['TP'])
+    fp_count = len(data['FP'])
+    fn_count = len(data['FN'])
+    return (tp_count, fp_count, fn_count)
+
+
+def test_collect_ner_results_both_empty():
+    utterance_targets = []
+    utterance_predictions = []
+    ner_filename = 'test.json'
+    collect_ner_results(utterance_targets, utterance_predictions, ner_filename)
+    (tp, fp, fn) = analyze_ner_file(ner_filename)
+    os.remove(ner_filename)
+    assert(tp == 0)
+    assert(fp == 0)
+    assert(fn == 0)
+
+
+def test_collect_ner_results_both_nonempty_equal():
+    utterance_targets = [[{'start': '10', 'end': '15', 'value': 'Chris',
+                           'entity': 'Person'}]]
+    utterance_predictions = [[{'start': '10', 'end': '15', 'value': 'Chris',
+                               'entity': 'Person'}]]
+    ner_filename = 'test.json'
+    collect_ner_results(utterance_targets, utterance_predictions, ner_filename)
+    (tp, fp, fn) = analyze__ner_file(ner_filename)
+    os.remove(ner_filename)
+    assert(tp == 1)
+    assert(fp == 0)
+    assert(fn == 0)
